refactor-repository-layer

+++ OrderSystem2/forms/farmer/FarmerDetail.cs
@@ -1,5 +1,6 @@
-﻿using OrderSystem2.forms.farmer;
-using OrderSystem2.forms.product;
+﻿using System.Diagnostics;
+using System.Text.RegularExpressions;
+using OrderSystem2.forms.farmer;
 using OrderSystem2.model;
 using OrderSystem2.repository.concretes;
 using OrderSystem2.service.concretes;
@@ -8,13 +9,9 @@ namespace OrderSystem2.forms
 {
     public partial class FarmerDetail : Form
     {
-        private string _connectionString = "Server=localhost;Database=OrderSystem;Integrated Security=True;TrustServerCertificate=True";
-
         private FarmerRepository _farmerRepository;
         private FarmerService _farmerService;
 
-        private ZoneRepository _zoneRepository;
-        private ZoneService _zoneService;
 
         private Farmer _farmer;
 
@@ -24,19 +21,35 @@ namespace OrderSystem2.forms
         private bool isDragging = false;
         private Point startPoint = new Point(0, 0);
 
+        private Dictionary<TextBox, string> initialValues;
+
         public FarmerDetail(Farmer farmer)
         {
             InitializeComponent();
             _farmer = farmer;
 
-            _farmerRepository = new FarmerRepository(_connectionString);
+            _farmerRepository = new FarmerRepository();
             _farmerService = new FarmerService(_farmerRepository);
 
-            _zoneRepository = new ZoneRepository(_connectionString);
-            _zoneService = new ZoneService(_zoneRepository);
-         
             LoadData();
-            Bind();
+            AttachPanelDragEvents(panel2);
+
+            buttonSave.Click -= buttonSave_Click;
+            buttonSave.Click += buttonSave_Click;
+
+            buttonDelete.Click -= buttonDelete_Click;
+            buttonDelete.Click += buttonDelete_Click;
+
+            buttonSave.Enabled = false;
+
+            foreach (var control in Controls.OfType<TextBox>())
+            {
+                control.TextChanged -= MarkAsChanged;
+                control.TextChanged += MarkAsChanged;
+            }
+
+            SaveInitialValues();
+
 
         }
 
@@ -44,41 +57,79 @@ namespace OrderSystem2.forms
         {
             textBoxFarmerID.Text = _farmer.Id.ToString();
             textBoxFarmerID.ReadOnly = true;
-            textBoxZoneId.Text = _zoneService.GetZoneNameByFarmerId(_farmer.Id);
-            textBoxZoneId.ReadOnly = true;
             textBoxName.Text = _farmer.Name;
             textBoxSurname.Text = _farmer.Surname;
             textBoxPhone.Text = _farmer.Phone;
             textBoxEmail.Text = _farmer.Email;
             textBoxAddress.Text = _farmer.Address;
             textBoxTc.Text = _farmer.Tc;
-            textBoxCreatedBy.Text = _farmer.CreatedBy.ToString();
-            textBoxCreatedBy.ReadOnly = true;
-            textBoxCreatedAt.Text = _farmer.CreatedAt.ToString();
-            textBoxCreatedAt.ReadOnly = true;
-            textBoxUpdatedBy.Text = _farmer.UpdatedBy.ToString();
-            textBoxUpdatedBy.ReadOnly = true;
-            textBoxupdatedAt.Text = _farmer.UpdatedAt.ToString();
-            textBoxUpdatedBy.ReadOnly = true;
+
         }
 
-        private void Bind()
+        private void SaveInitialValues()
         {
-            this.MouseDown += new MouseEventHandler(MouseDownHandler);
-            this.MouseMove += new MouseEventHandler(MouseMoveHandler);
-            this.MouseUp += new MouseEventHandler(MouseUpHandler);
+            initialValues = Controls.OfType<TextBox>()
+                   .ToDictionary(tb => tb, tb => tb.Text);
+            Debug.WriteLine("Initial Values Saved: ");
 
-
-            foreach (Control ctrl in this.Controls)
+            foreach (var control in this.Controls)
             {
-                ctrl.MouseDown += new MouseEventHandler(MouseDownHandler);
-                ctrl.MouseMove += new MouseEventHandler(MouseMoveHandler);
-                ctrl.MouseUp += new MouseEventHandler(MouseUpHandler);
+                if (control is TextBox textBox)
+                {
+                    Debug.WriteLine($"Saving: {textBox.Name} = {textBox.Text}");
+                }
             }
+        }
+
+        private bool HasChanges()
+        {
+            return Controls.OfType<TextBox>().Any(tb => tb.Text != initialValues[tb]);
+        }
+
+        private bool AreAllFieldsFilled()
+        {
+            return Controls.OfType<TextBox>().All(tb => !string.IsNullOrWhiteSpace(tb.Text));
+        }
+
+        private void MarkAsChanged(object sender, EventArgs e)
+        {
+            buttonSave.Enabled = HasChanges();
+        }
+
+        private bool IsValidEmail(string email)
+        {
+            string pattern = @"^[^@\s]+@[^@\s]+\.[^@\s]+$";
+            return Regex.IsMatch(email, pattern);
+        }
+
+        private bool IsValidPhoneNumber(string phoneNumber)
+        {
+            string pattern = @"^(\+90|90)?(5\d{9}|[2-9]\d{9})$";
+            return Regex.IsMatch(phoneNumber, pattern);
         }
 
         private void buttonSave_Click(object sender, EventArgs e)
         {
+            if (!AreAllFieldsFilled())
+            {
+                MessageBox.Show("Lütfen tüm alanları doldurun!", "Uyarı", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
+            }
+            if (textBoxTc.Text.Length != 11)
+            {
+                MessageBox.Show("Lütfen Tc'yi doğru giriniz!", "Hata", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return; 
+            }
+            if (!IsValidEmail(textBoxEmail.Text))
+            {
+                MessageBox.Show("Lütfen geçerli bir e-posta adresi girin!", "Hata", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
+            }
+            if (!IsValidPhoneNumber(textBoxPhone.Text))
+            {
+                MessageBox.Show("Lütfen geçerli bir telefon numarası girin! (Örn: 5321234567 veya 2123456789)", "Hata", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                return;
+            }
 
             try
             {
@@ -92,6 +143,8 @@ namespace OrderSystem2.forms
                 _farmerService.Update(_farmer, _farmer.Id);
 
                 MessageBox.Show("Güncelleme başarılı!", "Başarılı", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                buttonSave.Enabled = false;
+
                 LoadData();
 
                 FarmerForm farmerForm = Application.OpenForms.OfType<FarmerForm>().FirstOrDefault();
@@ -117,30 +170,44 @@ namespace OrderSystem2.forms
 
             if (confirmResult == DialogResult.Yes)
             {
-                try
+                bool hasAnyOrder = _farmerService.HasAnyOrders(_farmer.Id);
+                if (hasAnyOrder)
                 {
-                    _farmerService.Delete(_farmer.Id);
-
-                    MessageBox.Show("Kayıt başarıyla silindi!", "Başarılı", MessageBoxButtons.OK, MessageBoxIcon.Information);
-
-                    FarmerForm farmerForm = Application.OpenForms.OfType<FarmerForm>().FirstOrDefault();
-                    if (farmerForm != null)
+                    MessageBox.Show("Çiftçi silinemiyor. Sistemde işlemi bulunmakta.", "Başarılı", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    return;
+                }
+                else
+                {
+                    try
                     {
-                        farmerForm.LoadFarmer();
-                    }
+                        _farmerService.Delete(_farmer.Id);
 
-                    this.Close();
+                        MessageBox.Show("Kayıt başarıyla silindi!", "Başarılı", MessageBoxButtons.OK, MessageBoxIcon.Information);
+
+                        FarmerForm farmerForm = Application.OpenForms.OfType<FarmerForm>().FirstOrDefault();
+                        if (farmerForm != null)
+                        {
+                            farmerForm.LoadFarmer();
+                        }
+
+                        this.Close();
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show($"Hata: {ex.Message}", "Hata", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
                 }
-                catch (Exception ex)
-                {
-                    MessageBox.Show($"Hata: {ex.Message}", "Hata", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                }
+
             }
 
         }
-
         private void pictureBoxClose_Click(object sender, EventArgs e)
         {
+            FarmerForm farmerForm = Application.OpenForms.OfType<FarmerForm>().FirstOrDefault();
+            if (farmerForm != null)
+            {
+                farmerForm.LoadFarmer();
+            }
             this.Close();
         }
 
@@ -188,27 +255,36 @@ namespace OrderSystem2.forms
             this.Close();
         }
 
-        private void MouseDownHandler(object sender, MouseEventArgs e)
+        private void AttachPanelDragEvents(Panel panel)
+        {
+            panel.MouseDown += Panel_MouseDown;
+            panel.MouseMove += Panel_MouseMove;
+            panel.MouseUp += Panel_MouseUp;
+        }
+
+        private void Panel_MouseDown(object sender, MouseEventArgs e)
         {
             if (e.Button == MouseButtons.Left)
             {
                 isDragging = true;
-                startPoint = new Point(e.X, e.Y);
+                startPoint = e.Location;
             }
         }
 
-        private void MouseMoveHandler(object sender, MouseEventArgs e)
+        private void Panel_MouseMove(object sender, MouseEventArgs e)
         {
             if (isDragging)
             {
-                Point currentScreenPos = PointToScreen(e.Location);
+                Point currentScreenPos = ((Control)sender).PointToScreen(e.Location);
                 this.Location = new Point(currentScreenPos.X - startPoint.X, currentScreenPos.Y - startPoint.Y);
             }
         }
 
-        private void MouseUpHandler(object sender, MouseEventArgs e)
+        private void Panel_MouseUp(object sender, MouseEventArgs e)
         {
             isDragging = false;
         }
+
+
     }
 }
